Create new c cmd
